navigationsystem splited into components
