+++ front-end/src/components/navigationsystem/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./NavigationSystem";
+export type { PatientData, NavigationSystemProps } from "./types";
